requirements.txt
pdfminer.six >= 20200726
python-pptx >= 0.6.18
docx2txt >= 0.8